refactor(bytecode): move TextError to text.rs; drop dead validate_all (#770 review)
Review follow-ups scoped to this PR's own surface (nothing in the #776 array
zone): TextError now lives in isa::text next to its only producer (restoring
module cohesion; the always-0 line note travels with it), and the unused
validate_all — which also wrongly omitted simulate_stack — is removed. The
structure+arch composition it tested is now covered by a test exercising the
validate_structure+validate chain consumers actually use.
