+++ crates/bloqade-lanes-bytecode-core/src/isa/program.rs
@@ -103,35 +103,6 @@ impl std::fmt::Display for BinaryError {
 
 impl std::error::Error for BinaryError {}
 
-/// Error from text (`.sst`) parsing.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum TextError {
-    /// No `version M.N;` header was found before the first instruction.
-    MissingVersion,
-    /// The version header's value could not be parsed.
-    /// Currently unreachable: a malformed version fails the whole parse via `BadInstruction`.
-    /// Retained for API stability and potential future use.
-    InvalidVersion { line: usize, value: String },
-    /// A line could not be parsed as an instruction.
-    BadInstruction { line: usize, text: String },
-}
-
-impl std::fmt::Display for TextError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            TextError::MissingVersion => write!(f, "missing version header"),
-            TextError::InvalidVersion { line, value } => {
-                write!(f, "line {line}: invalid version '{value}'")
-            }
-            TextError::BadInstruction { line, text } => {
-                write!(f, "line {line}: cannot parse instruction '{text}'")
-            }
-        }
-    }
-}
-
-impl std::error::Error for TextError {}
-
 /// Serialize to the native binary format (see module docs).
 pub fn to_binary(program: &Program) -> Vec<u8> {
     let mut buf = Vec::with_capacity(HEADER_LEN + program.code.len() * INSTRUCTION_WIDTH as usize);
